feat: refactor to address comments
