+++ src/features/create-dao/Steps/DAOSummary.module.scss
@@ -1,0 +1,45 @@
+.Row {
+	display: flex;
+	margin-top: 0.625rem;
+	width: 100%;
+
+	.Column {
+		display: flex;
+		flex-direction: column;
+
+		.Heading {
+			margin-top: 0;
+		}
+
+		.MediaInput {
+			height: 12.5rem;
+			width: 12.5rem;
+			margin-right: 2rem;
+			border-radius: 0.5rem !important;
+		}
+	}
+
+	.SummaryRow {
+		gap: 4rem;
+		
+		.SummaryColumn {
+			gap: 1rem;
+		}
+	}
+}
+
+.Confirmation {
+	color: #F58C61;
+	text-align: center;
+	border-top: solid 0.063rem #232323;
+	padding: 1.5rem 0;
+	line-height: 1.5rem;
+
+	.ConfirmationQuestion {
+		font-weight: 700;
+	}
+}
+
+.Footer {
+	margin-top: 1rem;
+}
